Break out the `convert` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`convert` command.
